Change statistics HTML attachment base path to '/government/statistics'
Statistics are available at "/government/statistics/:slug", but associated
HTMLAttachments are given the base path "/government/publication/:slug/:attachment".

This is due to the HTMLAttachment class using the Attachable's class name to form
the url, and OfficialStatistics and NationalStatistics are subtypes of Publication.

This PR adds an `html_attachment_type` method to Publication, allowing the class
to differentiate between different subtypes and allow HTMLAttachment to
set the  basepath to "/government/statistics".

If this method is not available to a class, HTMLAttachment will continue to
fallback to the Attachable's class name.