Unpublish and redirect to external URLs
A number of live documents have been requested to be redirected to
various external pages, hosted on etl.beis.gov.uk and
https://www.judiciary.uk.

As these pages are live on GOV.UK and in Whitehall these documents need
to be unpublished and consolidated into alternative urls rather than
just a simple redirect being applied to keep Whitehall in sync with the
rest of our publishing stack.

Something to note is that this code is currently designed to delete any
drafts that exist for the URL's which is a prerequisite to unpublishing
a document.

Trello:
https://trello.com/c/cjn3BZKZ/2350-2-redirect-beis-and-hmcts-urls