Fix up withdrawn documents which erroneously have draft editions
For years, we've been incorrectly saving a draft edition in
Publishing API after informing it of Withdrawn documents in
Whitehall. This was fixed in #5992, but we still have a number of
existing documents this affects which would currently be unable
to have their withdrawal explanations updated.

This commit adds a rake task to resync all withdrawn documents
with Publishing API, explicitly removing any draft editions that
exist. This should solve the issue once and for all.

I've had to change the `PublishingApiUnpublishingWorker` to
inherit from `WorkerBase` in order to get the
`perform_async_in_queue` method. I see no disadvantage to doing
so, and it should free up the main queue for more important
publishing.

I also had to remove some lines from `ResluggingTest`, which are
no longer needed now that we're loading all Rake tasks in
test_helper.rb.

Trello: https://trello.com/c/Oicf9IVG/2354-5-investigate-fix-withdrawn-explanations-not-updating-in-whitehall