Fail gracefully when no draft exists in Publishing API
We're running this code against all withdrawn editions that could
have been affected by bulk publishing in the past. We don't know
for sure which ones are affected, without doing an audit in
Whitehall and in Publishing API.

When we attempt to remove a draft for a withdrawn edition, and
find that no draft exists, then that's actually the end state we
want already, so isn't an issue. We want to handle the 404 error
gracefully and move on to the next withdrawn document to process.
Depending on the state of the document, Publishing API may also
return a 422 response, so we catch both here. See
https://github.com/alphagov/publishing-api/blob/8b9b753866a013d012093f93a51c0809e2a74f96/app/commands/v2/discard_draft.rb#L27

This commit was added after running this on integration and
encountering the issue:

```
12:59:49 Fixing edition 602f7faa-7631-11e4-a3cb-005056011aef, locale en
12:59:49 Fixing edition 602f5e6c-7631-11e4-a3cb-005056011aef, locale en
12:59:49 Fixing edition 602f6556-7631-11e4-a3cb-005056011aef, locale en
12:59:49 Fixing edition 5ef0254b-7631-11e4-a3cb-005056011aef, locale en
12:59:49 Fixing edition e01ab0bc-f479-4b1e-a2ca-70b9bd70aa31, locale en
12:59:50 rake aborted!
12:59:50 GdsApi::HTTPNotFound: URL: https://publishing-api.integration.govuk-internal.digital/v2/content/e01ab0bc-f479-4b1e-a2ca-70b9bd70aa31/discard-draft
12:59:50 Response body:
12:59:50 {"error":{"code":404,"message":"There is not a draft edition of this document to discard"}}
```