Add more tests to publishing_api_test.rb
Now that we've run the rake task added in #5995, we no longer need
the task - or its associated tests - and will remove them in the
next commit.

However, we don't want to lose the precedent we've set of
beginning to test our rake tasks. It will be less of a barrier for
future developers to write test-driven rake tasks if there exists
a framework within which to write the tests. For that reason, I've
written some tests for the `publishing_api:bulk_republish:for_organisation`
rake task, which will mean publishing_api_test.rb itself won't
be deleted in the next commit.

The hope is that future rake tasks will have tests written for
them, and perhaps the existing rake tasks will have some more
tests retrospectively written for them like in this commit.

Trello: https://trello.com/c/Oicf9IVG/2354-5-investigate-fix-withdrawn-explanations-not-updating-in-whitehall