Add healthcheck endpoint for scheduled editions without a job
This adds an endpoint `/healthcheck/unenqueued_scheduled_editions` to
check for the number of scheduled editions that have not been queued in
Sidekiq.

We already have this data included in the `/healthcheck` endpoint, but
this is shown in Icinga for every host (rather than once per app). Once
this change has been deployed and the alert set up, the old system can
be removed.