Fix code Rubocop crashes on
Rubocop has a cop that checks for shuffle instances that could be
replaced with sample. The cop crashes with this code as it doesn't
anticipate that an endless range could be provided.

The code we have here is pretty bizarre, it shuffles and an array and
then returns all but the first element. This seems a totally strange use
case and I felt it was better to change this code rather than resolve
the bug in Rubocop.