Add edition attachments to fact check request
Editions sent to fact check only displayed the main document
and any inline attachments.

This change adds the attachment partial for editions with
attachments, so that all attachments can be fact-checked.

This appears on both the `edit` page where the user can
submit the fact check, and the `show` page where the user
is thanked for submitting the check and can re-review the
document.