Do not include plurals when comparing locale files
Different languages require different keys for creating plural forms of
words. The tests we have currently assume all languages work in the same
way as English, which is an incorrect assumption.

Therefore removing the check on plural keys. Later work will take place
to remove these tests and replace them with a gem that can perform the
task more elegantly.