Adds temp rake task to migrate Whitehall docs
Migration of a number of documents from GDS to CDDO. The attached CSV
was provided externally and contains the slug, lead orgs, supporting
orgs and 'document type' for each document. The ordering of the
organisations is important as it dictates the order in the frontend and
so this has been taken into consideration.

The rake task turned out to be a little bespoke for the given CSV
unfortunately due to a few different edge cases but as it's only a
temporary task I think this is fine.

Trello:
https://trello.com/c/pdTmWCJ2/2464-content-tagging-for-gds-mog-change-cddo-whitehall-content