Add and configure 'i18n-coverage' gem
This is the gem used by Content Publisher to track which i18n keys
are untested (and therefore possibly unused). The coverage report
is generated in `coverage/i18n`, which is git-ignored.

Introducing the dependency causes some tests to fail, which is a
fairly common side-effect of i18n-coverage on GOV.UK [1] [2], so
is wrapped in a conditional so that it is only included when we
want to execute it.

[1]: https://github.com/alphagov/collections/pull/2370/commits/11ff0507c70e86a0e1e5017201966f9b2cb5dbdd
[2]: https://github.com/alphagov/govuk_publishing_components/pull/2038/commits/cd3b54b81c2cf8d954df240b831b8127f7579134