Run i18n-tasks 'normalize' task across locale files
This sorts keys alphabetically, removes stray newlines, and
generally improves our YAML structure. This gets run implicitly
when running any other i18n-tasks command, and as we plan to use
i18n-tasks to remove unused keys, we want to minimise the diff to
review when we start to do that over the next few commits.