Do not display document type for Topical Events
In e7192cf2e4, the conditional for Topical Events was removed. This
previously had the effect of not displaying the document type for
Topical Events. As there is no published edition for these document
types, the `display_type_key` method fails due to a NilClass and a 500
error is returned.

Therefore adding the conditional back in so we don't attempt to show a
document type for Topical Events.