Move and rewrite testing docs into README
I've kept the section about Mustache templates separate, but above
the testing section so it's seen first, although the commands in
there should be covered as standard by GOV.UK Docker setup.

I don't know enough about Mustache to extend the docs, but recall
getting some errors for it in the past, so I think it's helpful to
at least introduce it in the README.

The script to run JavaScript tests simply started a Rails Server in
a test environment, on a different port. From my own testing, the
environment doesn't matter, so I've removed the script in favour of
reusing or starting the server normally.