Update check_organisation_links_worker.rb
Bad binary expression operand order
In a binary expression, the literal should preferably be on the right side of the expression. Yoda expressions, which are binary expressions spelled in reverse to how we normally speak, affect code readability & understandibility. This can be fixed by simply shifting the literal to the right and the variable to the left.