Fix failing tests
These tests are failing because there is another validation causing the 
the flash[:alert] to not match completely e.g. 

```
expected:
-"There is already an active draft edition for this document"
actual:
+"There is already an active draft edition for this document and Edition 
organisations edition can't be blank"
```

This seems to be as a result of the rails 6.1 validations. Since the 
other message is not relevant to this test it seems ok to match to the 
substring in this instance.