Add warning to dangerous rake task
We ran this rake task in production, and it stuck 31,000 items on
publishing_api's downstream_high queue. Effectively blocking all
publishing for over an hour.

Until we've fixed the task so that it uses only low-priority queues, or
until we've come up with a less naïve way of republishing HTML
attachments let's put a warning in the task description.