Always present telephone numbers as left-to-right
When a telephone number is included in the contact details of an
organisation page which is translated in a right-to-left language, we
are incorrectly applying this to the phone number.

This resulted in `+972 (0) 2 541 4100` being presented as `4100 541 2
(0) 972+`, which is incorrect, even if read right-to-left.

By always including the telephone number as left-to-right, we will
ensure the correct number is displayed.