Silence deprecation warning
Chaining multiple scopes that use the same column will be deprecated in Rails 6.2.
We do this on whitehall a lot because there is a default scope on the Edition model.
Unscoping the default before we apply a new scope removes the warning, but it
causes a large number of tests to fail. Rather than introduce a big code change in
this PR, silencing the warning and will address issue in subsequent work.