Skip S3 healthcheck if being run from E2E tests
TLDR:

This commit skips the S3 healthcheck if the app is being run from the
E2E tests.

More info:

We added a S3 healthcheck [1] for Whitehall which seems to have resulted
in Whitehall not being able to setup properly, which has then broke the
e2e tests when initialising Whitehall's docker container.

This situation occurs as a consequence of provisioning the E2E Whitehall
container [2]. During that process the healthcheck endpoint is queried
[3], and dependant on the output (e.g. 200 vs 500) either flags the
container as healthy, or unhealthy [4]. If unhealthy, there is some
logic that waits a set amount of time (3 minutes) continually checking
for a healthy container [5] before continuing (due to apps being
dependant on Whitehall running [6]). After the wait was up, the Jenkins
job failed with:

```
12:23:15  bundle exec rake docker:wait_for_whitehall_admin
12:26:21  rake aborted!
12:26:21  Container(s) whitehall-admin were unhealthy after 60 seconds
```

Digging into this, the only actual error log I could find was on the CI
machine itself which recorded that the '/healthcheck/ready' endpoint
returned a 500. I assume this is because E2E tests can't talk to S3 and
thus, failed the healthcheck. Skipping the S3 healthcheck is fine for
this case as the E2E tests don't test any flows that interact with S3.

[1]: https://github.com/alphagov/whitehall/pull/6221
[2]: https://github.com/alphagov/publishing-e2e-tests/blob/master/Makefile#L93-L94
[3]: https://github.com/alphagov/whitehall/blob/main/Dockerfile#L32
[4]: https://github.com/alphagov/publishing-e2e-tests/blob/526b4b2edd88a6eeb6634d35c7186e51f2b265ca/lib/docker_service.rb#L66-L70
[5]: https://github.com/alphagov/publishing-e2e-tests/blob/526b4b2edd88a6eeb6634d35c7186e51f2b265ca/lib/docker_service.rb#L5-L15
[6]: https://github.com/alphagov/publishing-e2e-tests/blob/master/Makefile#L121-L122