Update style for non-Whitehall content group
There is currently a details disclosure element with a summary on the
"Add document to a group" view.

The styling currently means there is no pointer cursor, so users are not
aware this can be clicked on to expand the form.

Adding a pointer cursor to the summary and also adding the arrow
alongside to make it clearer that this element opens on click.