Update Design
- Group via century 
- Apply Design System
- Use Components 
- Add lazy loading to images
- Separate data and add to partials
- Remove redundant CSS

- Update a selection of headings link to the particular profile of that person. 
As headings are not typically link-able (and not within the heading component) this "custom" to this page and maintains previous functionality

- Adjusting CSS style naming to be more generic to apply wider.

- Due to additional work on this PR[1] lazy loading has been added to the image_cards to improve the performance.

- "extra_links" that is text only is also in use on the image_card 

- Add aria to ensure the markup reads as grouped lists to AT yet maintain the Design System layout and styling without adding additional code. (This also potentially allows for more fine grain control of what is within grouped lists)

- No alt on image as decorative

---
Explored using ol > li markup however decided against it due to the nested grouping that was needed to apply the Design System's rows and columns. 

This page is somewhat un-usual, in that this is list of images.  For example to convert to a list and use the grid system -  markup similar to this is needed which is invalid:

ol (21st century) > div (rows wrapper) > li (column)
---

[1]https://github.com/alphagov/govuk_publishing_components/pull/2270
Remove alt text from image