Fix missing Prime Minister on historic PM page
An update to govuk_publishing_components changed the parameter names in
Release 27.0.0. We use the image cards to show rows of previous PMs
here.

The incorrect parameter names caused the image card to not render when
an href was not passed.

By renaming the params, the page renders correctly again.