Update past foreign secretaries page
Following the Sept 2021 reshuffle, we have an end date for the first
entry on the page.

It was also decided that, for consisentency, this page, past chancellors
and past prime ministers should only deal with past holders of the
position.

In this case a foreign secretary was listed as incumbant and with a link
to their biography.

We have decided current holders of the post should only be represented
on pages about current government.