Migrate featured links into translations table
This will allow us to use the `globalize` gem to manage the translations
of featured links.

As the application code is deployed at the same time as this migration,
the FeaturedLink has already been translated at this point, so we can no
longer access url and title on the model, as Rails will query the
featured_link_translations table instead. Therefore using raw SQL to
gather the existing featured_links records for migration.