Unpublish withdrawn docs with redirect
These pages have been withdrawn, but the publisher actually wants them
to be unpublished with a redirect as their withdrawn status is causing
confusion. The migration re-publishes each document, and then
unpublishes them with a redirect. Doing the migration this way keeps
Whitehall in sync with the rest of the stack.

Zendesk: https://govuk.zendesk.com/agent/tickets/4724480