Add PolicyGroupDependency
This will be used to store dependencies (e.g. contacts) for policy
groups, in a similar way to EditionDependency.

By doing this, we can republish the policy group when any dependencies
are updated.

As part of this migration, we need to populate the table with any
existing dependencies, which can be done by parsing the Govspeak in each
policy group's body (which is stored in the `desription` field).