Republish dependent policy groups when contact is updated
When a contact has been changed, we need to update any policy groups
that use the information from within this contact.

Therefore adding a `after_update` callback to ensure dependent policy
groups are republished when a contact is updated.

There is no `after_destroy` callback as the policy group won't republish
correctly as the contact no longer exists. In this case, we will just
leave the contact embedded as it is. When the policy group is next
edited manually, the user will receive an error as the contact can no
longer be found.