Add help text for organisation details page
Unless the user has read the guidance, it is not clear how to update
the 'Description' field  in the 'Details' tab. It's not shown when the
user clicks "edit" on the Details tab - which is what users expect.

This should hopefully make it clearer and reduce the number of tickets
coming through to 2nd line.