Remove `latin_script?`
I believe this code was originally added to prevent documents for
certain languages (Arabic, Yiddish etc) assigning the documents slug the
same text as the title. Using text from these languages in the URL can
cause encoding errors.

The intention was to use the document id, instead of the title, for the
slug, if the `latin_script?` tag was false. However, this code is never
actually used and I don't think the issue is still present. I've tested
the behaviour in Arabic for a World News Article, and it successfully
publishes with the document id as the slug.