Fix new corporate information pages not showing up on "about us"
The "about us" corporate information page is special in that it
displays a list of an organisation's other corporate information pages
at the bottom.  This list comes from the links hash, which has the
content IDs of all the other pages to show.

If a new page is created, the about page needs republishing, so that
the content ID of the new page will be in its links hash.  Currently
this doesn't happen, so new corporate information pages don't show up
on the about page until it gets updated.

As a simple fix, just republish the about page when another page is
updated.  After deploying this, we should republish all the about
pages (there's a rake task for this), to ensure we don't have any
missing links.