Add worker to update coronavirus_travel db fields
I'm not sure if this is needed, but it adds a job to update the Rag
status for a world location on the date/time of "next_rag_applies_at".
Hopefully this means that when the api is called after the next rag
status has taken effect that manual intervention is not required to
updated the api data. This should also remove the need for any date
validation code in caller to determine whether or not to handle any
"the rag status is changing" actions.