Remove FakeRummagerApiForStatisticsAnnouncements
We have a class named DevelopmentModeStubs::FakeRummagerApiForStatisticsAnnouncements
which is used to mock Search API (previously known as Rummager) in
tests.

However it appears this mock is not being used for anything in any of
the tests (except the test for the mock itself), so it can be removed.

The mock also contains SQL queries that are not compatible with MySQL 8,
so are blocking us from upgrading.