Use traditional SQL mode
MySQL 5.7 changes the SQL mode to `ONLY_FULL_GROUP_BY`, however there
are multiple complex queries in this application that have been designed
for `TRADITIONAL` mode.

Therefore, in order to unblock the upgrade from MySQL 5.6 to 8, we
should explicitly switch to `TRADITIONAL` SQL mode so our existing
queries continue to work as designed.