Use MySQL in 'traditional' mode
MySQL 5.7.5 introduced new behaviour around its [handling of `GROUP BY` clauses][1]. This change resulted in some previously valid SQL queries in Whitehall to be deemed invalid and trigger MySQL errors. These errors were evident as failures in the app's test suite.

This change in behaviour can be reverted by removing `ONLY_FULL_GROUP_BY` from the `sql_mode` MySQL variable. It can also be achieved by setting the `sql_mode` to `TRADITIONAL`, which is a special '[combination SQL mode][2]' that excludes `ONLY_FULL_GROUP_BY`.

A reasonable amount of effort was put into refactoring the problematic SQL queries in the application to avoid needing to deviate from MySQL's now-default behaviour. However due to the complexity of the queries and their surrounding code, it was deemed more pragmatic to simply set the SQL mode to 'traditional'.

This commit therefore configures MySQL to operate in 'traditional' mode so it behaves as expected by the app.

[1]: https://dev.mysql.com/doc/refman/5.7/en/group-by-handling.html
[2]: https://dev.mysql.com/doc/refman/8.0/en/sql-mode.html#sql-mode-combo