Add test for invalid world location email signup
We currently have a test for an invalid email signup URL, but not for an
invalid World Location email signup URL (which is handled differently).

There have recently been multiple errors logged in Sentry for World
Location email signups, so adding this test will allow us to debug the
problem.

At present, this test will fail, so a fix will be implemented in a later
commit.