Fix return in email signup controller
Where a user attempts to signup to an invalid World Location (e.g.
because they have manually changed the URL), we return a not found
response.

However because the `return` was missing from this line, we then also
attempted to redirect them. This redirect called a method that then
attempted to retrieve a content ID from a non-existent World Location.

Adding the `return` here solves that problem and makes the test added in
the previous commit now pass.