Adds host to unpublishing redirect allow list
Content request came in from Highways England to enable unpublishing and
redirecting to their website for a document, due to a withdrawn
page causing confusion which needs to be unpublished with a redirect.

This isn't standard practice to redirect to an external host but has
been approved on our end.

Also moves the list of external hosts to a constant as this feels more
appropriate.

Zendesk: https://govuk.zendesk.com/agent/tickets/4803998