Updating Dockerfile with mutistage build and streamlining.
This commit streamlines the Whitehall image by installing dependencies etc in a new intermediate `builder` stage, before copying back the build outputs to the output image.

Adding a non-root user to run the application and own app-specific files.

A few  values have been removed such as, Healthcheck,PORT and DATABASE_URL - These are not needed as part of the Build stage.

Adding Puma as part of the commit, given that we want to run Puma in production, we are instead configuring it via govuk_app_config

Enable JSON-style logging has been removed from production.rb, this seems to be old code and unused. Looking in Integration Whiteall EC2 this log is empty.

https://trello.com/c/ZCL2pDyO/739-package-whitehall