Work around a MySQL 8.0.27 bug that broke sorting
There's [a bug in MySQL 8.0.27][1] which increases the memory required to order result sets containing JSON or TEXT fields.

This affected us when trying to list Policy Groups in the Whitehall Admin.

When a user visits:

```
/government/admin/groups
```

Rails performs a query like:

```
> SELECT * FROM policy_groups ORDER BY name;
ERROR 1038 (HY001): Out of sort memory, consider increasing server sort buffer size
```

Which was resulting in a 500 Internal Server Error response from the server.

---

MySQL query performance can decrease as the [sort buffer size][2] is increased, so it's important not to increase it more than is necessary.

[Guidance on the MySQL bug][3] suggests the sort buffer size needs to be at least 15 times that of the largest row in the sort. In our case, the longest value in the `policy_groups` table is 26227 characters (in the `description` TEXT field). Assuming this field makes up the bulk of the row's size, it means the sort buffer size should be at least 15 × 26227 = 393405.

The default value of 262144 is less than this. But doubling it to 524288 allows the query to execute successfully and seems to provide enough headroom for some flexibility.

---

This bug should be resolved in the next patch release of MySQL (version 8.0.28). We should therefore try removing this workaround once we're able to run against that newer version.

[1]: https://bugs.mysql.com/bug.php?id=105304
[2]: https://dev.mysql.com/doc/refman/8.0/en/server-system-variables.html#sysvar_sort_buffer_size
[3]: https://bugs.mysql.com/bug.php?id=105304#c517545