Align translation files with pluralisation rules
Where the pluralisation rule says only plural should exist, and the
singular and plural are identical, the singular has been removed; if
this is incorrect for a given language it would require a translator to
fix anyway.

However where the singular and plural forms are different, to err on the
side of caution the pluralisation rules have been updated rather than
deleting translations in order to make the tests pass.

For Vietnamese, though most of the singulars are identical to the
plurals, a small number are different; so again, I've chosen to err on
the side of caution.

For some blank translations, it's also been necessary to add some extra
blank entries (for languages that have more than just `:one` and
`:other` forms).

A small number of singulars have been deleted that appeared to be either
errors, or differing only by capitalisation from the plural.