Require new permissions to edit Coronavirus page
Adds new "Coronavirus Travel Editor" permission.
Makes this permission required to do view/edit etc the Coronavirus
travel page.

It's easier to do this as in this spike the coronavirus travel page has
a separate controller.

If users don't have this permission they are shown the standard
whitehall "permission denied" page.

Follows the same pattern as the Fatality Notices controller.