Redirect content to UKRI
UKRI (UK Research & Innovation) have asked us to redirect/withdraw a 
number of Whitehall documents in bulk, away from GOV.UK to their own 
website ukri.org.

We have created 2 tempory rake tasks one that simply unpublishes the 
content and another that unpublishes and redirects the content to the 
pages on the ukri site. 
This is based on the work done in: 
https://github.com/alphagov/whitehall/pull/5985
Trello card: 
https://trello.com/c/XbGfF8Ua/123-bulk-redirect-content-for-ukri