Reduce default cache max age to 5 minutes
Implements RFC 144 [1], by reducing cache-control max-age directive
to 5 minutes for GOV.UK pages served by Whitehall.

This does not honour the Cache Control set by Content Store. This
was a pragmatic decision since it would take a fair amount of effort
to update Whitehall and we are aiming to sunset whitehall-frontnd,
so instead we've updated the configuration to match the configuration
set in Puppet.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-144-consistent-cache-expiry.md