Add link to accessible format request form for participating organisations
A pilot scheme for a limited number of organisations is soon to run,
allowing users to submit accessible format requests via a form in
the Feedback app rather than sending an email.

This commit adds a link to attachments owned by organisations in the
pilot, replacing the current toggled "send an email" text.

To enable this, all Editions, Consultations and Responses now have access
to an organisation_in_accessible_format_request_pilot?. This defaults to nil
in most cases, but editions with eligable attachments will have access to
the alternative_format_provider where the method can identify if the owning
organisation is taking part in the pilot.

Throughout the new method is added as a parameter to the block_attachments method
which govspeak_helper uses to render the html. The _attachment partial then has access
to organisation_in_accessible_format_request_pilot? in order to render the link as needed.