Remove limit on Topical Event Features
Currently there is a limit(5) on the number of Features that are show for a Topical Event.
This was originally 6 - set here: https://github.com/alphagov/whitehall/commit/b81f8fece4b80141740bc5aeab0f31fabb25834e
and changed to 5 here: https://github.com/alphagov/whitehall/commit/aea2e46d1717db802543fbb69d5a77fee192facd.

We would like to show more Features for a Topical Event, and as there does not appear
to be a reason given as to why it was originally limited or why that limit was changed.
And as there appears to be no restriction on the number the can be added to a Topical
Event - it seems sensible to simply remove the limit.