Send an edition's auth_bypass_id with a whitehall asset when it gets created in asset manager.
This forms part of the work to implement shareable previews for whitehall editions. It ensures that draft content can be previewed with a preview token containing an auth_bypass_id rather than a signon account. It is implemented for the following assets:

* Images
* Consultation Response Forms
* File attachments

It does not apply to logos as these are not attached to editionable content.