Install Jasmine for Javascript tests
This is the first step in porting the JS tests from a qunit based Rails
gem, teaspoon to Jasmine. This has been done to resolve a consistency
issue where GOV.UK apps use Jasmine for JS testing in Rails applications
[1] and modernise the tooling, as Teaspoon uses old dependencies (the
version of qunit bundled is from 2015).

The other motivating factor was discovering that Whitehall won't work
correctly with the current version of Teaspoon [2] and noticing the
project seemed to no longer be particularly active.

[1]: https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#testing-javascript-code
[2]: https://github.com/alphagov/whitehall/commit/58c2e66c0c81a6112844b72c1cc2dca20ddd6bcf