Rename and organise JS tests
This updates the naming of each of the JS files so they are compatible
with Jasmine conventions (.spec). It then puts each file in a directory
that corresponds with the file they are testing. E.g.
admin/autocomplete.js now has a test at
spec/javascripts/admin/autocomplete.spec.js