Refactor parallel worker handling
This test has been refactored to resolve an exception that occurs if you
run this test outside of a parallel test:

```
NameError: uninitialized class variable @@worker_number in LinkReporterCsvServiceTest
```

This switches the class variable to use a class method that'll default
to nil (hence an || 0). By using the eigenclass we can set an instance
variable at the class level and avoid the linting violation.