Add a rake command to republish all drafts
This extends the existing collection of Publishing API rake tasks to include one which will bulk republish all documents which have an active draft edition.

This will be needed once the auth_bypass_id fields have been backfilled for existing Edition records. The changes need to be sent to the Publishing API so that shared preview links can be supported for exising drafts.

Trello: https://trello.com/c/PAZNzo4z/382-ensure-existing-whitehall-draft-editions-are-stored-with-authbypassids